add card1
Here are the arguments you added to Container:

Apply a padding of 16 on all sides of the box. Flutter units are specified in logical pixels, which are like dp on Android.
Constrain the container’s size to a width of 350 and a height of 450.
Apply BoxDecoration. This describes how to draw a box.
In BoxDecoration, set up DecorationImage, which tells the box to paint an image.
Set which image to paint in the box using an AssetImage, an image found in the starter project assets.
Cover the entire box with that image.
Apply a corner radius of 10 to all sides of the container.